refactor(course): change naming
