+++ src/features/courses/components/cards/course-card.tsx
