+++ src/features/courses/components/cards/instructor-card.tsx
