+++ src/features/courses/components/course-skeleton.tsx
