+++ src/features/courses/index.tsx
@@ -8,10 +8,14 @@ import Option from '@mui/joy/Option';
 
 import { SearchRounded } from "@mui/icons-material";
 
-import CourseList from "./components/courseList";
+import { useGetCoursesQuery } from "app/api/courses.api.slice";
+
+import CoursesList from "./components/courses-list";
 import BreadcrumbsCustom from "shared/components/breadcrumbsCustom";
 
 export default function Courses() {
+  const { data, isLoading } = useGetCoursesQuery("Course");
+
   return (
     <Box sx={{ flex: 1, width: '100%' }}>
       <Box sx={{ px: { xs: 2, md: 6 } }}>
@@ -55,7 +59,7 @@ export default function Courses() {
         </Select>
       </Sheet>
       <Divider sx={{ my: 2 }} />
-      <CourseList />
+      <CoursesList data={data} isLoading={isLoading} />
     </Box>
   )
 }
